requirements.txt
Django>=4.2
transformers
torch
sentencepiece
PyPDF2
python-docx
